thermal: fix parsing of spaced value attributes
In case a key=value pair has a value with a space in between the current
parser complain since it does not expect them.
This little patch make sure to append all the tokens without a key to the
previous key. This should to properly parse events such as:

  comm=AsyncTask #2 pid=6163 ratio=875

where the value "AsyncTask #2" should be mapped on the "comm" key.

[dieegg01; added initialization of prev_key to None]
[dieegg01; added test case TestSpacedValueAttributes::test_spaced_value_attr]

Change-Id: I287b662b8357bee9ac40d3e656b28e0deb3b31ff
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Juri Lelli <juri.lelli@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>