Add a method to save the image for the graphs
As a feature request by Juri
PlotLayout also handles the excess whitespaces in
images with boundary condition checks that are
added

In order to handle white spaces preveiously the
PlotLayout was adding an extra row for single
graphs so that the axes can be indexed in a 2D array
This extra row does not show up in QT or ipython
frontends but shows up in saved images.

The plot layout now handles single plots differently
and the logic for getting an axes corresponding to an
index is encapsulated within the PlotLayout object so
that the DataPlotters need not bother about it.

The modulo arithmetic change ensures that an extra row
is not added and the columns are clipped to num_plots if
the user gives a larger per_line value

Change-Id: Ic972c750573ceacbd2c12a95648dd67449713637
Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>