run: remember the trace classes that were created
We have a number of loops that do:

for attr in self.class.iterkeys():
    getattr(self, attr).blah

and similar.  All because there is no way of getting all the trace
classes of the Run object.  Simplify the code by storing all this
classes in a trace_classes array.  I'm keeping the classes in the main
Run object because I think it's nice to do:

r = Run()
r.thermal

While we are at it, rename "classes" to "class_definitions" to clearly
distinguish between trace_classes and class_definitions.

Change-Id: Iddf87d4dec7de83d0654af7519376e21dacabd91