base: get the maximum array length in a trace
Base.__get_trace_array_lengths() was calculating the size of the arrays
only for the first line of the trace, instead of parsing all the traces
as it did prior to ce7834839fad ("cr2/run,thermal: read trace file only
once").  Make it calculate the biggest size of all the arrays in the
trace.

Modify the asymmetric test to highlight the issue.

This has a 20% performance impact on a 28Mb thermal trace.  The impact
on traces without arrays should be minimal as we exit early it if
there are no arrays in the trace.

Change-Id: I47ad69512866e4d6690150e69b74525dfbf05af1