tests: bring testing trace.dat back to life
Testing trace.dat was removed in fa532d24c8fb ("tests: don't depend on
trace-cmd being installed").  Patrick is still invoking cr2 and
expecting it to generate the text version of the trace, so bring the
test code back in.  Untested code is dead code.

Skip the tests if trace-cmd is not installed.  This avoids the reason
why it was removed in the first place.

Change-Id: I900b24a3d41694451a089fb694119abdc41e3945