base: fix concatenation of numerical values
A previous patch:
   8deda27b thermal: fix parsing of spaced value attributes
introduced the support to parse key-value pairs where the value is a string
with spaces. This fix was mainly targeted to recover the complete task name
for "comm=" key which is reported by many event classes.
However, this patch is broken for other classes of events.

All the events in the sched_stat_* class report a unit of measure after
the numerical value of a key. These are some examples:

       sshd-1779  [000]    77.948169: sched_stat_sleep:     comm=sshd pid=1574 delay=210542666 [ns]
rcu_preempt-7     [000]    73.714509: sched_stat_runtime:   comm=rcu_preempt pid=7 runtime=184333 [ns] vruntime=17097006397 [ns]
     <idle>-0     [004]    67.138996: sched_stat_iowait:    comm=taskset pid=1773 delay=9385208 [ns]
     <idle>-0     [004]    67.128158: sched_stat_blocked:   comm=sh pid=1773 delay=5493958 [ns]

In general it make sense to concatenate a value to a following space separated
string only if the first token assigned to a key is itself a string.
In all other cases, i.e. the current key value is not a string, the
concatenation operation is not safe and also produce run-time exceptions.
This is the case of the previous trace events, where the a number is going
to be concatenated with a string.

This patch introduces a further check to enable concatention only if the
current key value is a string.
NOTE: in that case the current token is simply discarded, this should not be
      an issue since:
      1. the current token is not a "key=value" pair
      2. it is numerical value which could not be assigned to a
         corresponding key

This patch adds also a test case to verify proper parsing of string
concatenated values as well as numerical values (not concatenated to a
following string value).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>