tests: make test_parse_values_concatenation() register the sched_stat_runtime as part of the sched scope
When running nosetests from the tests/ directory, all the tests are
collated and run together, with only one import of cr2.  Because of
that, every register_dynamic() invocation is persistent and every test
run afterwards is affected by it.  As such, registering
sched_stat_runtime affects the result of test_get_filters() in
test_sched.py when nosetests is run from the tests/ directory.

Fix it by adding the appropriate scope to register_dynamic().

Change-Id: Ia8d87df4dce563f2838265c1fa3e695fd96ddcb9
Signed-off-by: Javi Merino <javi.merino@arm.com>