tests: make raw tests register sched_switch in scope "sched"
The addition of tests that dynamically register sched_switch have broken
an unrelated test:

======================================================================
FAIL: Test that Run::get_filters returns correct list of filters
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/javi/src/cr2/tests/test_sched.py", line 101, in test_get_filters
    self.assertTrue(len(sched_classes) == len(sched_filters))
AssertionError: False is not true

----------------------------------------------------------------------

This is because register_dynamic() is persistent, so this test is
affected by the dynamic registration of the sched_switch in the other
tests.  The "sched" filter makes the get_filters() count it as a sched
filter, but because the registration was done in the all class, it's not
picked up by run.sched_classes, so the test fails.

Change register_dynamic() invocations in the raw tests to set
scope="sched" for sched_switch to avoid the failure.

Change-Id: I1ae2ba459de1ffd01be603a7ad18c4db78fb808a
Signed-off-by: Javi Merino <javi.merino@arm.com>