doc/source: Add details about running WA on Linux
Mention in the documentation that Android SDK is optional for
users who plan to run WA on Linux devices only, and how they
would only be able to start running a limited number of workloads.

Also included a few trivial fixes such as spelling errors and
moving sentences around to improve flow.

Signed-off-by: Lisa Nguyen <lisa.nguyen@linaro.org>