run: update special fields regexp to parse task names with a space
Chrome spawn a number of tasks which have a space in between. This produces
trace files which could have events starting with, for example:

    AsyncTask #1-9241  [001]  1033.379578:

This patch update the regular expression used to parse the special fields
at the beginning of each event in such a way to properly grab the taskname
even when it presents a space in it.

The propose patch should be valid as long as we could assume there is not
tasks which name includes the "[" char.

This patch adds also a test case to verify proper parsing of all common
fields for a set of events. The set of events could easily be extended
to check for more tasks names.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>