run: don't fail if there is no GPU trace
Some old traces don't have GPU information and with the addition of
devfreq to the plots, they break summary_plots().  Let
get_all_freqs_data(), plot_freq_hists() and plot_load() cope with traces
without devfreq information.

Fix the number of freq plots while we are at it so that we only display
as many plots as we need and don't generate empty ones if the GPU
information is not there.

Change-Id: Id57eaf21771e711d72429ca38d74dbd2cc5eef5b
Signed-off-by: Javi Merino <javi.merino@arm.com>