cr2: move WA-specific results module to a cr2.wa namespace
Getting results is parsing a file that is generated by Workload
Automation.  Create a wa namespace where we can put all the WA-specific
functions of cr2.  Rename the CR2 class while we are at it, as this
class has not evolved as I envisioned it to and it's just a class that
encapsulates results.

Change-Id: Idead141ffdc47efa0b47d8d4273481aab32f3413