plotter: Let the ConstraintManager operate on all the permutations of multiple columns and runs
It can be useful to make multiple plots one for each run but each of
them showing multiple columns.  The ConstraintManager doesn't allow you
to populate the constraints like that and instead "zips" the runs and
columns.  Add an optional parameter to let it create constraints for all
the permutations of columns and runs.

Change-Id: Ib7ac20b5245c00f1ebf9883f8071b386db3fa6b9