run: sanitise textual traces generation
When we specify a target folder containing a trace data, the text trace is
generated only if missing while if already present we do not check for it
to be more recent than the corresponding binary trace.
This could produce some not noticeable errors when running multiple time
an experiment which override the same binary trace.

This patch refactors the trace generation code to grant that a "text trace
is always newer than the corresponding binary trace.

All the conditions of the previous code are covered but one: if a text
trace is present and newer than the binary one, we do not force the
generation of a "raw" trace in case it is missing.
Indeed, "raw" traces are generated contextually to "text ones, thus this
code seems to be useful only to revamp legacy experiments.
These old archives can still be update by simply "touching" a binary
trace before building a run object.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>