target: fix 'bl' module initialisation
When "load_default_modules" is set to False at target initialisation time
but the "bl" module is explicitly required, the initialization of this module
fails because the system as not yet been identified as a big.LITTLE.
This identification is performed by platform.update_from_target.

This patch post-pone modules initialisation at the end of the connection
method to ensure that all the required information are available at modules
probe time.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>