setup: Depend on trappy >= 2.0
trappy has changed the API for EventPlot in f4eefc02239a ("plotter:
EventPlot: Allow specification of lane names") which will be released
in trappy 2.0.  This is fixed in 5b1275f59fec ("sched: Update usage of
EventPlot API") in bart.  Make sure bart the next version of bart
depends on the appropriate version of trappy.