sched: Use Median for calculation of the period of a task
Here is what I do for calculating the period of the task:

Let's say a task started execution at the following times

                       T_1, T_2, ...T_n
Currently for align = "start" the period is

Average of:

                (T_2 - T_1), (T_4 - T_3), ....(T_n - T_{n-1})
In this method we have an edge condition when a task does not run
for a particular time or when it migrates between CPUs causing the
average to end up significantly higher than the actual period of
the task. However, as periodicity is statistical in nature
and can vary during the lifetime of the task, we can have a
good understanding of the periodicity by getting the median period
of the deltas above.

Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>