libs/wlgen/rta: fix calibration configuration using FIFO:0 by default
The configuration of ulimits for RT tasks could limit, on some systems,
the creation of RT tasks with a max priority of 0.
For calibration proposes this value could be in general acceptable.

TODO: we should read from target what is the maximum RT prio which can
be require, using "ultimi -Hr" and use that value to configure the
priority for the RT-App.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>