tools/scripts/cgroup: update to support CGroups with multiple controllers
If the same CGroup is created in multiple hierarchies (i.e. controllers)
this patch ensure that the workload will be allocated in all the same
CGroup for all controllers.

The lookup for a CGroup is done based on the CGroup name. Starting from
the CGroup mount point, all the folders which name matches the CGroup are
selected. The task PID is moved into each of these selected folders.

In case no CGroups matching the specified name can be found, than the script
returns with an error.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>