stats: Use super indexing for grammar
The current index does not unify indices from different DataFrames in a
statement it just picks up the index of the first object to be referenced
and reindexes the DataFrames encountered later in the statement using
the first encountered index.

This can lead to significant data loss if the initial dataframe has a
small or non overlapping index. In this patch, we unify the indices
before the reindexing happens.

Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>