run: don't expect classes to be global variables
Since 323bb8d22d25 ("run: register all classes using
Run.register_class()") all classes are registered through
Run.register_class().  There's no need to add them to the globals,
this is a legacy thing that we can now get rid of.

When registering a class, instead of recording it's __name__ in the
appropriate *_classes array, just store the class definition so that it
can be instanced when a Run() object is constructed.