sched: consistently name sched trace classes
When using Run(events=["cpu_capacity"]), we end up on re-using the
pre-defined sched::SchedCpuCapacity class, which is registered as
"sched_cpu_capacity", so it appears as run.sched_cpu_capacity.  This
misalignment is confusing for the user and should be avoided.

Rename all the pre-registered sched related events to match the
eventually user defined name.