ftrace: give more time for a trace to be collected
The time required to fetch a trace depends not only on it time length but
also on the amount of events collected and the processing power of the
target device. This patch tries to factorise these last two components
into a 5x constant which is good enough to collect a relatively big trace
on a relatively slow device.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>