systrace: identify CPUs count from sched_switch
SysTrace does not report the same medatata generated by trace-cmd
report. The "_cpus" attribute is required by plot_trace to properly
dimension the plot.

This patch ensures that this attribute is available by estimating the
number of CPUs from the events available in the trace.  The sched_switch
event is always enabled, thus it should be good enough to calculate the
number of cpus.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>