run: stop looping when we've reached the end of the window
In 0daba87249dd ("run: accept a window parameter"), we introduced
passing a window parameter to run.  __populate_data_from_line() raises
StopIteration when it sees a trace that crosses the end of the window,
with the intention... stop iterating.  Break the loop when we catch that
exception, which was the original intent instead of continue looping on
a trace that we know no more interesting lines.