tests/eas/rfc: refactor to support multiple wload configurations
RFC test cases share a common wload configuration syntax but could requires
different workloads. This patch refactor the RFC class to be a base class
which runs all the specified workload and a set of derived classes which
loads a specific workload configuration. For the time being we have
only the EAS configuration.

This should allows also to run only a subset of RFC tests if required
by naming the specific class to nosetests, e.g. this command:

  nosetests -v tests/eas/rfc.py:EAS

will run only EAS tests.


Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>