stats: Fix generation for trigger
The trigger.generate function needs to return the value
for the valid indices replaced with the trigger value
This value replacement was being handled in Aggregator which
was accessing a private member from the Trigger class.
Removing the _ for making it public was possibly a bad design
choice to make pylint happy.

Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>