tests: unregister dynamically created classes in the tests
The classes registered dynamically in the tests persist across the
execution of the testsuite.  That makes the testsuite sometimes fail if
a test has added a class to the sched scope and
test_ftrace_has_all_classes_scope_sched() run concurrently.  Unregister
the dynamic events when the tests that register them finish.