plot_utils: fix plot_allfreqs() for one actor
When the system has only one actor and you use trappy.summary_plots() or
trappy.compare_runs() you get:

    Traceback (most recent call last):
      File "trappy/tests/test_trappy.py", line 127, in test_summary_plots_one_actor
        trappy.summary_plots(self.actor_order, map_label)
      File "trappy/trappy/compare_runs.py", line 96, in summary_plots
        return compare_runs(actor_order, map_label, [(title, path)], **kwords)
      File "trappy/trappy/compare_runs.py", line 78, in compare_runs
        trappy.plot_utils.plot_allfreqs(run_data, map_label, **kwords)
      File "trappy/trappy/plot_utils.py", line 209, in plot_allfreqs
        run.plot_allfreqs(map_label, ax=ax)
      File "trappy/trappy/ftrace.py", line 531, in plot_allfreqs
        for this_ax, (label, dfr) in zip(ax, all_freqs):
    TypeError: zip argument #1 must support iteration

This is because plt.subplots() returns just the axis if ncols and nrows
are both 1, a 1D array if either ncols or nrows is 1 and a 2D array
otherwise.  plot_utils.plot_all_freqs() copes with ncols being 1 but not
with nrows being 1.  The latter happens when you use summary_plots() or
compare_runs() and there is only one actor.

Fix it so that axis is always a 2D array: the first index is the run and
the second the actor.