trappy: partially revert 8da875103915
In 8da875103915 ("ftrace: fix error with only one actor") we made
ftrace.plot_allfreqs() treat the case when there's only one actor a
special case that ignored the requirement of the "ax" parameter being an
array.  The documentation of plot_allfreqs() is correct: if ax is not
None, it must be an array of the same size as map_label.

Revert the change to ftrace.py and fix the test to match what the
documentation says.