lib/utils/trace{_analysis}: update events name to match new trappy naming
Without that patch the CPU and Cluster plots are not generated for traces
obtained using more recent versions of Trappy.
Indeed, Trappy is now using "standard" names for all events available in a
mainline kernel. This patch update the CPU frequency event to match the
canonical name of that event, i.e. "cpu_frequency".

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>