ipynb: sched_tune: add smoke test for a ramp task
This patch provides a smoke test notebook which plots data generated using
the RFC:STune test. To produce data suitable for this notebook:
1. update target.config to properly configure the target to test
2. run the sched tune smoke test:
   $ DEVMODE=1 source init_env
   $ nosetests -v tests/eas/rfc.py:STune
3. run the notebook provided by this patch

This notebook assume that the data are available in results_latest, i.e. that
the notebook is executed right after the completion of an RFC:STune test.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>