tests/eas/rfc: switch to usage of generic CGroups support provided by devlib
The configuration of the target for a test could require to setup some
CGroups controllers other than SchedTune. This patch adds the required
support to mount and configure each of the CGroup controllers specified
by the test configuration which are available in the target.

An example target configuration to defined two different CGroups setup is:

   /* Set of platform configurations to test */
   "confs" : [
       {
           "tag" : "noboost",
           "cgroups" : {
               "conf" : {
                       "schedtune" : {
                           "/"      : {"boost" :  0 },
                           "/stune" : {"boost" :  0 }
                       }
               },
               "default" : "/"
           }
       },
       {
           "tag" : "boost15",
           "cgroups" : {
               "conf" : {
                   "schedtune" : {
                       "/"      : {"boost" :  0 },
                       "/stune" : {"boost" : 15 }
                   },
                   "cpuset" : {
                       "/stune" : {"cpus" : "3" },
                       "/stune" : {"mems" : "0" }
                   }
               },
               "default" : "/stune"
           }
       }
   ],

In this setup, the "noboost" configuration setup the schedtune.boost attribute
for both the "/" (i.e. root) and the "/stune" CGrooups to be 0.
The second configuration, tagged "boost15" configure the schedtune.boost
attribute to be 0 (respectively 15) for the "/" (respectively "stune") CGroups,
as well as enforcing tasks in the "/stune" group to run only on CPU3.

The "default" attribute defined in which CGroup workload's tasks must be
moved for their execution.

TODO: add the support to map for each task a specific CGroup on which CGroup
      they should be executed.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>