Refactored _plot (now _resolve) in StaticPlot
Refactored the function to make it shorter and more readible. Also now making
it now possible to create a list of tuples containing all related information
e.g. the data to plot, axis and title. This could be used to handle plots in
a subclass and will hopefully make code more reusable with interactive plots.