libs/utils/env: fix topology building for ARM SMP platforms
On SMP ARM platforms, there is not distinction between these two attributes:
   devlib::target::big_core
   devlib::target::little_core
which breaks the code for the identification of clusters.

This patch ensures that, in case an ARM platform is not big.LITTLE, we can
still build a clusters representation which is build considering just the
cluster ID. This is the same approach generally used for x86_64 platforms.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>