libs/utils/env: report cluster frequencies for SMP platforms
For SMP platforms we are not currently reporting available frequencies
because the intial SMP support was targeting targets using the intel_pstate
driver.

This patch improves a little that situation by exploiting the "cpufreq"
module once available, which is usually the case for SMP ARM systems.
This patch reports a list of available frequencies for each cluster available
in the target system.

NOTE: if all clusters are in the same frequency domain, still we report
      a frequencies list for each cluster.

TODO: add a representation of frequency domains.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>