Added a BarPlot which can plot data as vertical bars
BarPlot added as a subclass of StaticPlot.

Values are plotted against their position in the list of data.

Series can be grouped(defualt) or stacked by setting the stacked
kwarg to True.

Spacing between groups is adjusted by the spacing kwarg. This is a
float. A proportion of the size of each group which sould be
spacing. e.g. 0.2 (defualt) means that 1/5 of the groups total
width is used as a spacing between groups.