tests/eas/rfc: add FLAGS support for each target configuration
This patch adds the support to have, for each target configuration, a new
"flags" attribute which can be used to fine tune the execution of the
workloads on that configurations.

The "flags" attribute is just a string which can contains one or more tokens.
Each token corresponds to a specific feature, for example:
 - ftrace: enable the collection of an FTrace, across the execution of the
           workload, using the common configuration specified in the
           configuration file

Other flags will be added to fine tune other features of an experiment
execution.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>