tests/eas/rfc: STune: update RAMP workload definition
Workloads which have a periodicity of 60Hz are quite common on mobile platforms
are represents some of the most critical tasks which impacts the user
experience, e.g. SurfaceFlinger in Android.

This patch configure the ramp workload to use a 16[ms] period, thus allowing
a better evaluation of SchedTune features on responsiveness of rendering tasks.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>