tests/eas/rfc: drop logging from test_execution_complete()
The logging interferes with the output from nosetests:

Check that data have been collected from the target ... 04:08:04  INFO    : Check for data being collected
ok

Drop the message as it is redundant.  Also drop the return as it's not
needed.  The test should be implemented using standard unittest.TestCase
assert methods (self.assert*), not by returning a value.  Put a TODO
placeholder instead.