libs/utils/energy.py: Initial integration of AEP Energy Meter:
This patch integrates the initial version of the ARM Energy Probe AEP
Energy Meter class.

The AEP delivers sample voltage, power and current values (10Khz). The
AEP Energy Meter uses only the power sample values and calculates the
energy value by multiplying the average power value with the time
difference value between the call of the reset() and the report()
method.

Current limitations:

 - The AEP Energy Meter class is tailored to measure DIE energy on the
   Hikey board for now.
 - It implements only the reset() and report()(sample()) method (e.g.
   used by the eas rfc test: tests/eas/rfc.py).
 - It uses only AEP channel 1 (hardcoded).
 - It uses a hardcoded value for shunt resistor (33 mOhm).
 - It uses 'LITTLE' channel tag to let the existing tooling work with
   the Hikey board even though we actually measure DIE energy.

The AEP Energy Meter operates on top of the devlib EnergyProbeInstrument
which itself uses the caiman binary to operate the AEP.

The EnergyProbeInstrument figures out if the caiman binary is not
installed and issues an appropriate error message in this case on the
host.

This patch bases on an initial version by Leo Yan <leo.yan@linaro.org>.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>