target: warn if a module is not supported for the target
When you create a target using devlib, you may not set
load_default_modules because you want to load a specific list of
modules.  In that case, raise a warning if one of those modules failed
to load, as it is something you are not expecting.

In the general case, the standard set of modules are tried to be loaded.
We know that they are not supported for all platforms so just emit a
debug message.