libs/utils/results: make more generic the parsing of energy.json
When just one cluster is available the energy.json file does not
reports all the "big" and "LITTLE" channels and the current code
throw an exception.

This patch makes the code a bit more robust by assuming a 0W
energy when a channel is not reported in the json file.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>