Enable DEVMODE by default
The usage of local libraries is (for the time being) the most
used version of the suite. Indeed many features are provided by
out-of-mainline patches provided only by local libraries.

This patch switch to the usage of local libraries if not otherwise
specified. If you want to use system installed libraries just export
    DEVMODE=0
in the shell where init_env is sourced and/or the ipython server
started.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>